Flatten repo structure by moving server files up
- all logdetective/server modules are now in logdetective/
- all loggers (except contrib update) now use logdetective.config.LOG
- load_skip_snippet_patterns moved from utils.py to config.py
- all references properly renamed
- [server] subpackage removed from pyproject, server deps are now the
  part of the base package
