+++ files/run_server.sh
@@ -28,4 +28,4 @@ else
 fi
 
 echo "Starting application server..."
-exec python -m gunicorn -c "$conf" logdetective.server.server:app
+exec python -m gunicorn -c "$conf" logdetective.server:app
