+++ logdetective/database/base.py
