+++ logdetective/database/models/exceptions.py
