+++ logdetective/database/models/merge_request_jobs.py
@@ -16,8 +16,8 @@ from sqlalchemy import (
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.engine import Row
 
-from logdetective.server.database.base import Base, transaction
-from logdetective.server.utils import retry_database_error
+from logdetective.database.base import Base, transaction
+from logdetective.utils import retry_database_error
 
 
 if TYPE_CHECKING:
