--- logdetective/server/models.py
@@ -1,546 +1,0 @@
-import datetime
-import subprocess as sp
-from typing import List, Dict, Optional, Any, Union, Sequence
-from pydantic import (
-    BaseModel,
-    Field,
-    model_validator,
-    field_validator,
-    NonNegativeFloat,
-    HttpUrl,
-    ConfigDict,
-)
-
-from logdetective.constants import (
-    DEFAULT_TEMPERATURE,
-    SYSTEM_ROLE_DEFAULT,
-    DEFAULT_MAXIMUM_ARTIFACT_MIB,
-    MINIMUM_SNIPPET_TRUNCATION_LEN,
-)
-from logdetective.models import PromptConfig
-
-
-class ArtifactBase(BaseModel):
-    """Base build artifact model."""
-
-    model_config = ConfigDict(hide_input_in_errors=True, extra="forbid")
-
-    name: str = Field(min_length=1, max_length=255, pattern=r"^[a-zA-Z0-9._\-\/ ]+$")
-
-
-class ArtifactFile(ArtifactBase):
-    """Model of one artifact as raw data"""
-
-    content: str = Field(description="Artifact file content as a string")
-
-
-class RemoteArtifactFile(ArtifactBase):
-    """Model of artifact linked with URL. By default, request size is limited to 50 MiB.
-    This can affect what kind of artifacts can be submitted."""
-
-    url: HttpUrl = Field(description="URL of artifact.")
-
-
-class BuildMetadata(BaseModel):
-    """Model of additional information provided about the build."""
-
-    specfile: Optional[str] = Field(
-        description="Contents of package spec file as a string.", default=None
-    )
-    last_patch: Optional[str] = Field(
-        description="Contents of last patch applied as a string.", default=None
-    )
-    commentary: Optional[str] = Field(
-        description="Comment attached to the triggered build, such as PR description.",
-        default=None,
-    )
-    infra_status: Optional[str] = Field(
-        description="State of build infrastructure as a string.", default=None
-    )
-
-
-class AnalysisRequest(BaseModel):
-    """Model of the request body for /analyze endpoint"""
-
-    model_config = ConfigDict(hide_input_in_errors=True, extra="forbid")
-    files: Sequence[Union[ArtifactFile, RemoteArtifactFile]] = Field(
-        description="List of artifacts",
-        min_length=1,
-        max_length=15,
-    )
-    build_metadata: Optional[BuildMetadata] = Field(
-        description="Optional build metadata. Unused in the initial release.",
-        default=None,
-    )
-
-    @model_validator(mode="after")
-    def check_unique_filenames(self) -> "AnalysisRequest":
-        """Check that files do not have duplicate names."""
-        names = [f.name for f in (self.files or [])]
-        if len(names) != len(set(names)):
-            raise ValueError("Duplicate filenames detected in 'files' list")
-        return self
-
-
-class JobHook(BaseModel):
-    """Model of Job Hook events sent from GitLab.
-    Full details of the specification are available at
-    https://docs.gitlab.com/user/project/integrations/webhook_events/#job-events
-    This model implements only the fields that we care about. The webhook
-    sends many more fields that we will ignore."""
-
-    # The unique job ID on this GitLab instance.
-    build_id: int
-
-    # The identifier of the job. We only care about 'build_rpm' and
-    # 'build_centos_stream_rpm' jobs.
-    build_name: str = Field(pattern=r"^build.*rpm$")
-
-    # A string representing the job status. We only care about 'failed' jobs.
-    build_status: str = Field(pattern=r"^failed$")
-
-    # The kind of webhook message. We are only interested in 'build' messages
-    # which represents job tasks in a pipeline.
-    object_kind: str = Field(pattern=r"^build$")
-
-    # The unique ID of the enclosing pipeline on this GitLab instance.
-    pipeline_id: int
-
-    # The unique ID of the project triggering this event
-    project_id: int
-
-
-class EmojiMergeRequest(BaseModel):
-    """Model of the 'merge_request' subsection of Emoji webhook messages.
-    This model implements only the fields that we care about. The webhook
-    sends many more fields that we will ignore."""
-
-    # The identifier of the target project
-    target_project_id: int
-
-    # The internal identifier (relative to the target project)
-    iid: int
-
-
-class EmojiHook(BaseModel):
-    """Model of Job Hook events sent from GitLab.
-    Full details of the specification are available at
-    https://docs.gitlab.com/user/project/integrations/webhook_events/#job-events
-    This model implements only the fields that we care about. The webhook
-    sends many more fields that we will ignore."""
-
-    # The kind of webhook message. We are only interested in 'emoji' messages
-    # which represents awarding or revoking emoji reactions on notes.
-    object_kind: str = Field(pattern=r"^emoji$")
-
-    # Information about the merge request this emoji applies to, if any.
-    merge_request: EmojiMergeRequest = Field(default=None)
-
-
-class Explanation(BaseModel):
-    """Model of snippet or general artifact explanation from Log Detective"""
-
-    text: str
-
-    def __str__(self):
-        """Return text of the Explanation"""
-        return self.text
-
-
-class Solution(BaseModel):
-    """Proposed solution to the issue Log Detective found."""
-
-    text: str
-
-    def __str__(self):
-        """Return text of the Solution"""
-        return self.text
-
-
-class Snippet(BaseModel):
-    """Model for snippets created by extractors.
-
-    text: original snippet text
-    line_number: location of snippet in the original build artifact
-    source_file: name of the original build artifact
-    """
-
-    text: str
-    line_number: int
-    source_file: Optional[str]
-
-
-class AnalyzedSnippet(Snippet):
-    """Model for snippets already processed by Log Detective.
-
-    snippet_analysis: analysis from the LLM
-    """
-
-    snippet_analysis: str
-
-
-class AgentResponse(BaseModel):
-    """Model for the final agent response, containing explanation,
-    proposed solution and a flag indicating if any issue has been found.
-
-    explanation: Explanation
-    solution: Proposed solution to the detected issue
-    no_issue_found: Set to true if no issue was detected
-    """
-
-    explanation: Explanation
-    solution: Optional[Solution] = None
-    no_issue_found: bool = False
-
-
-class APIResponse(AgentResponse):
-    """Model of data returned by Log Detective API
-
-    explanation: Explanation
-    snippets: List of extracted snippets
-    solution: Proposed solution to the detected issue
-    no_issue_found: Set to true if no issue was detected
-    """
-
-    snippets: Optional[List[Union[AnalyzedSnippet, Snippet]]] = None
-
-
-class KojiResponse(BaseModel):
-    """Model of data returned by Log Detective API when called when a Koji build
-    analysis is requested. Contains list of reponses to prompts for individual
-    snippets.
-    """
-
-    task_id: int
-    log_file_name: str
-    response: APIResponse
-
-
-class InferenceConfig(BaseModel):  # pylint: disable=too-many-instance-attributes
-    """Model for inference configuration of logdetective server."""
-
-    max_tokens: int = -1
-    api_timeout: float = 30.0
-    # Format: "provider:model_id", e.g. "openai:gpt-4o", "vertexai:gemini-2.0-flash"
-    model: str = "openai:default-model"
-    temperature: NonNegativeFloat = DEFAULT_TEMPERATURE
-    system_role: str = SYSTEM_ROLE_DEFAULT
-    total_max_retries: int = 10
-    max_retries_per_step: int = 3
-    # Provider-specific settings passed to the beeai backend via the settings dict.
-    # Keys are beeai internal setting names (e.g. api_key, base_url, vertex_project, ...).
-    provider_settings: dict[str, str] = {}
-    # Retry settings for transient LLM API errors (timeouts, rate limits).
-    # Uses exponential backoff (with jitter). Set retry_max_tries to 1 to disable.
-    retry_max_tries: int = 3
-    retry_max_time: int = 120
-
-
-class ExtractorConfig(BaseModel):
-    """Model for extractor configuration of logdetective server."""
-
-    max_clusters: int = 8
-    verbose: bool = False
-    max_snippet_len: int = Field(
-        description="Maximum length of extracted snippets",
-        ge=MINIMUM_SNIPPET_TRUNCATION_LEN,
-        default=2000
-    )
-    csgrep: bool = False
-    csgrep_timeout: float = 1.0
-    python_traceback: bool = False
-
-    @field_validator("csgrep", mode="before")
-    @classmethod
-    def verify_csgrep(cls, v: bool):
-        """Verify presence of csgrep binary if csgrep extractor is requested."""
-        if not v:
-            return v
-        try:
-            result = sp.run(
-                ["csgrep", "--version"],
-                text=True,
-                check=True,
-                shell=False,
-                capture_output=True,
-                timeout=1.0,
-            )
-        except (FileNotFoundError, sp.TimeoutExpired, sp.CalledProcessError) as ex:
-            raise ValueError(f"Required binary `csgrep` was not found in path: {ex}") from ex
-        if result.returncode == 0:
-            return v
-        raise ValueError(f"Issue was encountered while calling `csgrep`: {result.stderr}")
-
-
-class GitLabInstanceConfig(BaseModel):  # pylint: disable=too-many-instance-attributes
-    """Model for GitLab configuration of logdetective server."""
-
-    name: str
-    url: str = "https://gitlab.com"
-    # Path to API of the gitlab instance, assuming `url` as prefix.
-    api_path: str = "/api/v4"
-    api_token: Optional[str] = None
-
-    # This is a list to support key rotation.
-    # When the key is being changed, we will add the new key as a new entry in
-    # the configuration and then remove the old key once all of the client
-    # webhook configurations have been updated.
-    # If this option is left empty or unspecified, all requests will be
-    # considered authorized.
-    webhook_secrets: Optional[List[str]] = None
-
-    timeout: float = 5.0
-
-    # Maximum size of artifacts.zip (default: 50 MiB)
-    # In config, the unit is in MiB, but this max_artifact_size attribute will be in bytes
-    max_artifact_size: int = DEFAULT_MAXIMUM_ARTIFACT_MIB * 1024**2
-
-    @field_validator("max_artifact_size", mode="before")
-    @classmethod
-    def megabytes_to_bytes(cls, v: Any):
-        """Convert max_artifact_size from megabytes to bytes."""
-        return (v if isinstance(v, int) else DEFAULT_MAXIMUM_ARTIFACT_MIB) * 1024**2
-
-
-class GitLabConfig(BaseModel):
-    """Model for GitLab configuration of logdetective server."""
-
-    instances: Dict[str, GitLabInstanceConfig] = {}
-
-    @model_validator(mode="before")
-    @classmethod
-    def set_gitlab_instance_configs(cls, data: Any):
-        """Initialize configuration for each GitLab instance"""
-        if not isinstance(data, dict):
-            return data
-
-        instances = {}
-        for instance_name, instance_data in data.items():
-            instance = GitLabInstanceConfig(name=instance_name, **instance_data)
-            instances[instance.url] = instance
-
-        return {"instances": instances}
-
-
-class KojiInstanceConfig(BaseModel):
-    """Model for Koji configuration of logdetective server."""
-
-    name: str = ""
-    xmlrpc_url: str = "https://koji.fedoraproject.org/kojihub"
-    tokens: List[str] = []
-
-
-class KojiConfig(BaseModel):
-    """Model for Koji configuration of logdetective server."""
-
-    instances: Dict[str, KojiInstanceConfig] = {}
-    analysis_timeout: int = 15
-
-    # in yaml config, this is given in MiB, but we use bytes in code (same as gitlab)
-    max_artifact_size: int = DEFAULT_MAXIMUM_ARTIFACT_MIB * 1024**2
-
-    @field_validator("max_artifact_size", mode="before")
-    @classmethod
-    def megabytes_to_bytes(cls, v: Any):
-        """Convert max_artifact_size from megabytes to bytes."""
-        return (v if isinstance(v, int) else DEFAULT_MAXIMUM_ARTIFACT_MIB) * 1024**2
-
-    @model_validator(mode="before")
-    @classmethod
-    def set_koji_instance_configs(cls, data: Any):
-        """Initialize configuration for each Koji instance."""
-        if isinstance(data, dict):
-            instances = {}
-            for instance_name, instance_data in data.get("instances", {}).items():
-                instances[instance_name] = KojiInstanceConfig(
-                    name=instance_name, **instance_data
-                )
-            data["instances"] = instances
-        return data
-
-
-class LogConfig(BaseModel):
-    """Logging configuration"""
-
-    name: str = "logdetective"
-    level_stream: str | int = "INFO"
-    level_file: str | int = "INFO"
-    path: str | None = None
-    format: str = "%(levelname)s:%(name)s:%(asctime)s:%(message)s"
-    datefmt: str = "%Y-%m-%d %H:%M:%S"
-
-
-class GeneralConfig(BaseModel):
-    """General config options for Log Detective"""
-
-    packages: List[str] = []
-    excluded_packages: List[str] = []
-    sentry_dsn: HttpUrl | None = None
-    collect_emojis_interval: int = 60 * 60  # seconds
-    # max_artifact_size in config.yml is in MiBs, here (GeneralConfig class) is in bytes
-    max_artifact_size: int = DEFAULT_MAXIMUM_ARTIFACT_MIB * 1024**2
-    block_localhost_urls: bool = True
-    generate_solution: bool = True
-    delay_artifact_download: bool = False
-    # Timeout for execution of analysis in seconds
-    agent_timeout: int = 600
-    annotation_lookup_tool: bool = False
-    max_annotations: int = 3
-
-    @field_validator("max_artifact_size", mode="before")
-    @classmethod
-    def megabytes_to_bytes(cls, v: Any):
-        """Convert max_artifact_size from megabytes to bytes."""
-        return (v if isinstance(v, int) else DEFAULT_MAXIMUM_ARTIFACT_MIB) * 1024**2
-
-
-class Config(BaseModel):
-    """Model for configuration of logdetective server."""
-
-    log: LogConfig = Field(default_factory=LogConfig)
-    inference: InferenceConfig = Field(default_factory=InferenceConfig)
-    # TODO(jpodivin): Extend to work with multiple extractor configs
-    extractor: ExtractorConfig = Field(default_factory=ExtractorConfig)
-    gitlab: GitLabConfig = Field(default_factory=GitLabConfig)
-    koji: KojiConfig = Field(default_factory=KojiConfig)
-    general: GeneralConfig = Field(default_factory=GeneralConfig)
-    prompts: PromptConfig = Field(default_factory=PromptConfig)
-
-    @field_validator("prompts", mode="before")
-    @classmethod
-    def make_prompts_optional(cls, v: Any) -> Any:
-        """Convert undefined into an empty dict."""
-        return {} if v is None else v
-
-
-class TimePeriod(BaseModel):
-    """Specification for a period of time.
-
-    If no indication is given
-    it falls back to a 2 days period of time.
-
-    Can't be smaller than a hour"""
-
-    weeks: Optional[int] = None
-    days: Optional[int] = None
-    hours: Optional[int] = None
-
-    @model_validator(mode="before")
-    @classmethod
-    def check_exclusive_fields(cls, data):
-        """Check that only one key between weeks, days and hours is defined,
-        if no period is specified, fall back to 2 days."""
-        if isinstance(data, dict):
-            how_many_fields = sum(
-                1
-                for field in ["weeks", "days", "hours"]
-                if field in data and data[field] is not None
-            )
-
-            if how_many_fields == 0:
-                data["days"] = 2  # by default fallback to a 2 days period
-
-            if how_many_fields > 1:
-                raise ValueError("Only one of months, weeks, days, or hours can be set")
-
-        return data
-
-    @field_validator("weeks", "days", "hours")
-    @classmethod
-    def check_positive(cls, v):
-        """Check that the given value is positive"""
-        if v is not None and v <= 0:
-            raise ValueError("Time period must be positive")
-        return v
-
-    def get_time_period(self) -> datetime.timedelta:
-        """Get the period of time represented by this input model.
-        Will default to 2 days, if no period is set.
-
-        Returns:
-            datetime.timedelta: The time period as a timedelta object.
-        """
-        delta = None
-        if self.weeks:
-            delta = datetime.timedelta(weeks=self.weeks)
-        elif self.days:
-            delta = datetime.timedelta(days=self.days)
-        elif self.hours:
-            delta = datetime.timedelta(hours=self.hours)
-        else:
-            delta = datetime.timedelta(days=2)
-        return delta
-
-    def get_period_start_time(
-        self, end_time: Optional[datetime.datetime] = None
-    ) -> datetime.datetime:
-        """Calculate the start time of this period based on the end time.
-
-        Args:
-            end_time (datetime.datetime, optional): The end time of the period.
-                Defaults to current UTC time if not provided.
-
-        Returns:
-            datetime.datetime: The start time of the period.
-        """
-        time = end_time or datetime.datetime.now(datetime.timezone.utc)
-        if time.tzinfo is None:
-            time = time.replace(tzinfo=datetime.timezone.utc)
-        return time - self.get_time_period()
-
-
-class MetricTimeSeries(BaseModel):
-    """Recorded values of given metric"""
-
-    metric: str
-    timestamps: List[datetime.datetime]
-    values: List[float]
-
-
-class MetricResponse(BaseModel):
-    """Requested metrics"""
-
-    time_series: List[MetricTimeSeries]
-
-
-class ContributionSnippet(BaseModel):
-    """Annotated log snippet from the user contribution to logdetective.com."""
-    text: str = Field(min_length=1)
-    user_comment: str = Field(min_length=1)
-
-
-class ContributionLog(BaseModel):
-    """Log file entry containing annotated snippets."""
-    snippets: list[ContributionSnippet] = []
-
-    @field_validator("snippets", mode="before")
-    @classmethod
-    def filter_invalid_snippets(cls, v):
-        """Keep only well-formed snippet dicts; silently drop the rest."""
-        if not isinstance(v, list):
-            return []
-        return [
-            i for i in v if isinstance(i, dict) and i.get("text") and i.get("user_comment")
-        ]
-
-
-class ContributionBuild(BaseModel):
-    """User contribution from logdetective.com with annotated build failure and snippets."""
-    fail_reason: str = Field(min_length=1)
-    how_to_fix: str = Field(min_length=1)
-    logs: dict[str, ContributionLog]
-
-    @field_validator("logs", mode="before")
-    @classmethod
-    def filter_invalid_log_entries(cls, v):
-        """Keep only valid log entries; silently drop the rest."""
-        if not isinstance(v, dict):
-            raise ValueError("logs must be a dict")
-        return {k: entry for k, entry in v.items() if isinstance(entry, dict)}
-
-    @model_validator(mode="after")
-    def check_has_valid_snippets(self) -> "ContributionBuild":
-        """We need at least 1 valid snippet in a contribution to be able to look up anything."""
-        if sum(len(entry.snippets) for entry in self.logs.values()) == 0:
-            raise ValueError("no valid snippets found")
-        return self
