+++ logdetective/templates/gitlab_full_comment.md.j2
