+++ logdetective/templates/gitlab_short_comment.md.j2
