+++ tests/server/test_server_analyze.py
@@ -4,8 +4,8 @@ import pytest
 import pytest_asyncio
 from httpx import ASGITransport, AsyncClient
 
-from logdetective.server.server import app, get_http_session, validate_request_size
-from logdetective.server.config import SERVER_CONFIG
+from logdetective.server import app, get_http_session, validate_request_size
+from logdetective.config import SERVER_CONFIG
 
 from tests.server.test_helpers import mock_AnalyzeRequestMetrics
