+++ tests/server/test_server_gitlab.py
@@ -14,19 +14,19 @@ from sqlalchemy import select
 from sqlalchemy.orm import selectinload
 from beeai_framework.backend import ChatModel
 
-from logdetective.server.gitlab import (
+from logdetective.gitlab import (
     is_eligible_package,
     retrieve_and_preprocess_koji_logs,
     check_artifacts_file_size,
 )
-from logdetective.server.gitlab import process_gitlab_job_event
-from logdetective.server.models import JobHook, GitLabInstanceConfig, APIResponse, Explanation
-from logdetective.server.database.models import (
+from logdetective.gitlab import process_gitlab_job_event
+from logdetective.models import JobHook, GitLabInstanceConfig, APIResponse, Explanation
+from logdetective.database.models import (
     AnalyzeRequestMetrics,
     Forge,
     GitlabMergeRequestJobs,
 )
-from logdetective.server.exceptions import LogsTooLargeError, LogDetectiveArtifactsMissingError
+from logdetective.exceptions import LogsTooLargeError, LogDetectiveArtifactsMissingError
 
 from tests.server.test_helpers import (
     DatabaseFactory,
@@ -260,7 +260,7 @@ def mock_analysis(mocker, request):
         snippets=[]
     )
     return mocker.patch(
-        "logdetective.server.gitlab.analyze_artifacts",
+        "logdetective.gitlab.analyze_artifacts",
         AsyncMock(return_value=mock_response)
     )
 
@@ -383,7 +383,7 @@ async def test_fallback_to_task_failed_log_if_no_match(
     zip_content = create_zip_archive(files)
     mock_artifact_download(mocker, zip_content)
     mocker.patch(
-        "logdetective.server.gitlab.check_artifacts_file_size", return_value=True
+        "logdetective.gitlab.check_artifacts_file_size", return_value=True
     )
 
     mock_session = mocker.AsyncMock()
@@ -408,7 +408,7 @@ async def test_raises_file_not_found_on_no_failures(
     zip_content = create_zip_archive(files)
     mock_artifact_download(mocker, zip_content)
     mocker.patch(
-        "logdetective.server.gitlab.check_artifacts_file_size", return_value=True
+        "logdetective.gitlab.check_artifacts_file_size", return_value=True
     )
 
     mock_session = mocker.AsyncMock()
@@ -427,7 +427,7 @@ async def test_raises_logs_too_large_error(mocker: MockerFixture, gitlab_cfg, mo
     Tests that LogsTooLargeError is raised if check_artifacts_file_size returns False.
     """
     mocker.patch(
-        "logdetective.server.gitlab.check_artifacts_file_size", return_value=False
+        "logdetective.gitlab.check_artifacts_file_size", return_value=False
     )
     mock_to_thread = mocker.patch("asyncio.to_thread")
     mock_session = mocker.AsyncMock()
@@ -524,7 +524,7 @@ async def test_architecture_prioritization(mocker: MockerFixture, gitlab_cfg, mo
     zip_content = create_zip_archive(files)
     mock_artifact_download(mocker, zip_content)
     mocker.patch(
-        "logdetective.server.gitlab.check_artifacts_file_size", return_value=True
+        "logdetective.gitlab.check_artifacts_file_size", return_value=True
     )
     mock_session = mocker.AsyncMock()
 
@@ -548,7 +548,7 @@ async def test_toplevel_failure_fallback(mocker: MockerFixture, gitlab_cfg, mock
     zip_content = create_zip_archive(files)
     mock_artifact_download(mocker, zip_content)
     mocker.patch(
-        "logdetective.server.gitlab.check_artifacts_file_size", return_value=True
+        "logdetective.gitlab.check_artifacts_file_size", return_value=True
     )
     mock_session = mocker.AsyncMock()
 
@@ -577,7 +577,7 @@ async def test_unrecognized_architecture_handling(
     zip_content = create_zip_archive(files)
     mock_artifact_download(mocker, zip_content)
     mocker.patch(
-        "logdetective.server.gitlab.check_artifacts_file_size", return_value=True
+        "logdetective.gitlab.check_artifacts_file_size", return_value=True
     )
     mock_session = mocker.AsyncMock()
