+++ tox.ini
@@ -99,4 +99,4 @@ per-file-ignores =
 description = Run djlint
 deps =
     djlint>=1.36.0
-commands = djlint logdetective/server/templates
+commands = djlint logdetective/templates
